add encrypt and decrypt subcommand handling
